Test listener code
This code tests that the endpoints in listener.py are set up correctly, and
parse the arguments in notifications correctly.  It doesn't run against a real
database---if we change the code to call 'find' or 'delete' in a slightly
different way, the tests may need to change.  (Most likely, we might start
checking that the resource mapping found when deleting has the correct
arguments.)